Convert site to SvelteKit
