+++ packages/site/package.json
@@ -4,10 +4,9 @@
 	"version": "0.0.1",
 	"type": "module",
 	"scripts": {
-		"dev": "sapper dev --ext '.svelte .svx'",
-		"build:site": "sapper build --legacy --ext '.svelte .svx'",
-		"export": "sapper export --legacy --ext '.svelte .svx'",
-		"start": "node __sapper__/build",
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
 		"cy:run": "cypress run",
 		"cy:open": "cypress open",
 		"test": "run-p --race dev cy:run",
@@ -15,11 +14,8 @@
 	},
 	"private": "true",
 	"dependencies": {
-		"compression": "^1.7.1",
 		"mdsvex": "^0.7.1-beta.3",
-		"polka": "^0.5.0",
 		"refractor": "^2.10.0",
-		"sirv": "^0.4.0",
 		"yootils": "^0.0.15"
 	},
 	"devDependencies": {
@@ -33,20 +29,22 @@
 		"@rollup/plugin-json": "^4.1.0",
 		"@rollup/plugin-node-resolve": "^13.0.5",
 		"@rollup/plugin-replace": "^3.0.0",
+		"@sveltejs/kit": "next",
+		"@sveltejs/adapter-auto": "next",
 		"codemirror": "^5.49.2",
 		"npm-run-all": "^4.1.5",
-		"prism-svelte": "^0.4.4",
+		"prism-svelte": "^0.5.0",
 		"rehype-autolink-headings": "^2.0.5",
 		"rehype-slug": "^2.0.3",
 		"remark-syntax-highlight": "^0.1.3",
 		"rollup": "^2.58.0",
 		"rollup-plugin-svelte": "^7.1.0",
 		"rollup-plugin-terser": "^7.0.2",
-		"sapper": "^0.29.3",
-		"svelte": "^3.7.1",
+		"svelte": "^3.49.0",
 		"svelte-json-tree": "^0.0.7",
 		"typeface-catamaran": "^0.0.72",
 		"typeface-roboto": "^0.0.75",
-		"unist-util-visit": "^2.0.1"
+		"unist-util-visit": "^2.0.1",
+		"vite": "^3.0.4"
 	}
 }
