+++ packages/site/src/app.html
@@ -4,8 +4,6 @@
 		<meta content="width=device-width, initial-scale=1.0" name="viewport" />
 		<meta charset="utf-8" />
 
-		%sapper.base%
-
 		<link rel="stylesheet" href="global.css" />
 		<link rel="manifest" href="manifest.json" />
 		<link rel="preload" as="image" href="/penguin.gif" />
@@ -48,14 +46,12 @@
 		/>
 		<meta property="twitter:image" content="https://mdsvex.com/og.png" />
 
-		%sapper.styles% %sapper.head%
+		%sveltekit.head%
 	</head>
 
 	<body>
-		<div id="sapper">%sapper.html%</div>
+		<div>%sveltekit.body%</div>
 
-		%sapper.scripts%
-    
     <!-- Cloudflare Web Analytics --><script defer src='https://static.cloudflareinsights.com/beacon.min.js' data-cf-beacon='{"token": "590d70c530b74f41b117351532dd6809"}'></script><!-- End Cloudflare Web Analytics -->
 	</body>
 </html>
