+++ packages/site/src/routes/docs.svelte
@@ -1,5 +1,5 @@
 <script>
-	import { stores } from '@sapper/app';
+	import { page } from '$app/stores';
 	import { onMount } from 'svelte';
 	import { fade } from 'svelte/transition';
 	import docs from './_docs.svtext';
@@ -10,8 +10,6 @@
 	let width = 1100;
 	let current;
 	let position = '';
-
-	const { page } = stores();
 
 	const nav = [
 		['Install', 'docs#install-it'],
